+++ xmovie/test/test_presets.py
@@ -1,4 +1,3 @@
-import cartopy.crs as ccrs
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 import numpy as np
@@ -6,10 +5,16 @@ import pytest
 import xarray as xr
 
 from xmovie.presets import (
+    _add_coast,
+    _add_land,
     _check_input,
     _core_plot,
     _smooth_boundary_NearsidePerspective,
+    rotating_globe,
+    rotating_globe_dark,
 )
+
+from . import has_cartopy, requires_cartopy
 
 
 def test_check_input():
@@ -36,17 +41,20 @@ def test_check_input():
 )
 def test_core_plot(plotmethod, expected_type, filled):
     da = xr.DataArray(np.random.rand(4, 6))
-    fig, ax = plt.subplots()
+    _, ax = plt.subplots()
     pp = _core_plot(ax, da, plotmethod=plotmethod)
     assert isinstance(pp, expected_type)
     if filled is not None:
         assert pp.filled == filled
 
 
+@requires_cartopy
 @pytest.mark.parametrize("lon", [-700, -300, 1, 300, 700])
 @pytest.mark.parametrize("lat", [-200, -90, 0, 90, 180])
 @pytest.mark.parametrize("sat_height", [35785831, 45785831])
 def test_smooth_boundary_NearsidePerspective(lon, lat, sat_height):
+    import cartopy.crs as ccrs
+
     lon = -100
     lat = -40
     sat_height = 35785831
@@ -56,5 +64,21 @@ def test_smooth_boundary_NearsidePerspective(lon, lat, sat_height):
     # modify the projection with smooth boundary
     pr_mod = _smooth_boundary_NearsidePerspective(pr)
 
+    assert type(pr) is type(pr_mod) and isinstance(pr_mod, ccrs.Projection)
     assert pr.proj4_params == pr_mod.proj4_params
     assert pr.globe == pr_mod.globe
+
+
+@pytest.mark.skipif(has_cartopy, reason="no req check error if cartopy is installed")
+@pytest.mark.parametrize("fn", [_add_coast, _add_land, rotating_globe, rotating_globe_dark])
+def test_cartopy_req_check(fn):
+    with pytest.raises(RuntimeError, match="Required modules failed to import: 'cartopy'"):
+        fn()
+
+
+@requires_cartopy
+@pytest.mark.parametrize("fn", [_add_coast, _add_land])
+def test_raise_no_geoax(fn):
+    _, ax = plt.subplots()
+    with pytest.raises(TypeError, match="^Cannot add land on non-cartopy axes."):
+        fn(ax, "standard")